+++ src/test/java/nextstep/courses/SessionPeriodTest.java
@@ -1,5 +1,6 @@
-package nextstep.qna.domain;
+package nextstep.courses;
 
+import nextstep.courses.domain.SessionPeriod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
